reorganization: creation of Enzo/cosmology directory
