+++ src/Enzo/CMakeLists.txt
@@ -48,6 +48,7 @@ target_link_options(enzo PRIVATE ${Cello_TARGET_LINK_OPTIONS})
 
 # add source files from various subdirectories to the enzo target
 add_subdirectory(amr)
+add_subdirectory(cosmology)
 add_subdirectory(feedback-particle)
 add_subdirectory(gravity)
 add_subdirectory(hydro-mhd) # add files in hydro-mhd to enzo
